+++ cli/commands/pack.go
@@ -1,18 +1,26 @@
 package commands
 
 import (
+	"fmt"
 	"os"
+	"path/filepath"
 	"strings"
 
 	"github.com/apex/log"
 	"github.com/spf13/cobra"
 
+	"github.com/tarantool/cartridge-cli/cli/common"
+	"github.com/tarantool/cartridge-cli/cli/context"
 	"github.com/tarantool/cartridge-cli/cli/pack"
 )
 
 var (
 	packTypeArgs = []string{"tgz", "rpm", "deb", "docker"}
+	deps         = []string{}
+	depsFile     = ""
 )
+
+const defaultPackageDepsFile = "package-deps.txt"
 
 func init() {
 	rootCmd.AddCommand(packCmd)
@@ -41,6 +49,89 @@ func init() {
 	packCmd.Flags().StringVar(
 		&ctx.Pack.StatboardUnitTemplatePath, "stateboard-unit-template", "", stateboardUnitTemplateUsage,
 	)
+
+	packCmd.Flags().StringSliceVar(&deps, "deps", []string{}, depsUsage)
+	packCmd.Flags().StringVar(&depsFile, "deps-file", "", depsFileUsage)
+}
+
+func addTarantoolDepIfNeeded(ctx *context.Ctx) error {
+	if ctx.Tarantool.TarantoolIsEnterprise || !(ctx.Pack.Type == pack.RpmType || ctx.Pack.Type == pack.DebType) {
+		return nil
+	}
+
+	var tarantoolVersion string
+	if ctx.Pack.Type == pack.RpmType {
+		tarantoolVersion = strings.SplitN(ctx.Tarantool.TarantoolVersion, "-", 2)[0]
+	} else if ctx.Pack.Type == pack.DebType {
+		tarantoolVersion = ctx.Tarantool.TarantoolVersion
+	}
+
+	if err := ctx.Pack.Deps.AddTarantool(tarantoolVersion); err != nil {
+		return fmt.Errorf("Failed to get tarantool dependency: %s", err)
+	}
+
+	return nil
+}
+
+func parsePackageDeps(deps []string, depsFile string) (common.PackDependencies, error) {
+	var err error
+
+	if depsFile != "" && len(deps) != 0 {
+		return nil, fmt.Errorf("You can't specify --deps and --deps-file flags at the same time")
+	}
+
+	if depsFile == "" && len(deps) == 0 {
+		defaultPackeDepsFilePath := filepath.Join(ctx.Project.Path, defaultPackageDepsFile)
+		if _, err := os.Stat(defaultPackeDepsFilePath); err != nil {
+			return nil, fmt.Errorf("Failed to use default package dependencies file: %s", err)
+		}
+
+		depsFile = defaultPackeDepsFilePath
+	}
+
+	if depsFile != "" {
+		if _, err := os.Stat(depsFile); os.IsNotExist(err) {
+			return nil, fmt.Errorf("Invalid path to file with dependencies: %s", err)
+		}
+
+		content, err := common.GetFileContent(depsFile)
+		if err != nil {
+			return nil, fmt.Errorf("Failed to get file content: %s", err)
+		}
+
+		deps = strings.Split(content, "\n")
+	}
+
+	parsedDeps, err := common.ParseDependencies(deps)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to parse dependencies file: %s", err)
+	}
+
+	return parsedDeps, nil
+}
+
+func fillDependencies(ctx *context.Ctx) error {
+	if ctx.Pack.Type == pack.RpmType || ctx.Pack.Type == pack.DebType {
+		var err error
+
+		if ctx.Pack.Deps, err = parsePackageDeps(deps, depsFile); err != nil {
+			return err
+		}
+
+		return addTarantoolDepIfNeeded(ctx)
+	}
+
+	if depsFile != "" || len(deps) != 0 {
+		flagName := "deps"
+		if depsFile != "" {
+			flagName = "deps-file"
+		}
+
+		log.Warnf("You specified the --%s flag, but you are not packaging RPM or DEB. "+
+			"Flag will be ignored", flagName)
+	}
+
+	return nil
 }
 
 var packCmd = &cobra.Command{
@@ -75,6 +166,10 @@ func runPackCommand(cmd *cobra.Command, args []string) error {
 	}
 
 	if err := pack.FillCtx(&ctx); err != nil {
+		return err
+	}
+
+	if err := fillDependencies(&ctx); err != nil {
 		return err
 	}