+++ cli/commands/usage.go
@@ -55,6 +55,12 @@ on creation build and runtime images`
 defaults to "TARANTOOL_SDK_PATH" env`
 
 	sdkLocalUsage = `Deliver the SDK from the local machine`
+
+	depsUsage = `Dependencies for the RPM and DEB packages`
+
+	depsFileUsage = `Path to the file that contains dependencies
+for the RPM and DEB packages.
+`
 )
 
 // RUNNING
@@ -85,7 +91,7 @@ defaults to ./instances.yml ("cfg" in .cartridge.yml)`
 
 	daemonizeUsage = `Start instance(s) in background`
 
-	stateboardUsage = `Manage application stateboard as well as instances 
+	stateboardUsage = `Manage application stateboard as well as instances
 ("stateboard" in .cartridge.yml)`
 
 	stateboardOnlyUsage = `Manage only application stateboard`