+++ cli/replicasets/failover_priority.go
@@ -19,7 +19,7 @@ func SetFailoverPriority(ctx *context.Ctx, args []string) error {
 		return err
 	}
 
-	if ctx.Replicasets.FailoverPriorityNames, err = common.GetInstancesFromArgs(args, ctx); err != nil {
+	if ctx.Replicasets.FailoverPriorityNames, err = common.GetInstancesFromArgs(args, ctx.Project.Name); err != nil {
 		return err
 	}