+++ test/integration/pack/test_pack.py
@@ -15,13 +15,14 @@ from utils import assert_filemodes
 from utils import assert_files_mode_and_owner_rpm
 from utils import validate_version_file
 from utils import check_package_files
-from utils import assert_tarantool_dependency_deb
-from utils import assert_tarantool_dependency_rpm
+from utils import assert_tarantool_dependency_deb, assert_dependencies_deb
+from utils import assert_tarantool_dependency_rpm, assert_dependencies_rpm
 from utils import run_command_and_get_output
 from utils import build_image
 from utils import get_rockspec_path
+from utils import tarantool_version
 
-from project import set_and_return_whoami_on_build
+from project import set_and_return_whoami_on_build, replace_project_file
 
 
 # ########
@@ -103,6 +104,17 @@ def custom_base_image(session_tmpdir, request, docker_client):
         f.write("FROM centos:8")
 
     build_image(session_tmpdir, 'my-custom-centos-8')
+
+
+@pytest.fixture(scope="session")
+def tarantool_versions():
+    min_deb_version = re.findall(r'\d+\.\d+\.\d+-\d+-\S+', tarantool_version())[0]
+    max_deb_version = str(int(re.findall(r'\d+', tarantool_version())[0]) + 1)
+    min_rpm_version = re.findall(r'\d+\.\d+\.\d+', tarantool_version())[0]
+    max_rpm_version = max_deb_version  # Their format is the same
+
+    return {"min": {"deb": min_deb_version, "rpm": min_rpm_version},
+            "max": {"deb": max_deb_version, "rpm": max_rpm_version}}
 
 
 # ########
@@ -990,3 +1002,230 @@ def test_verbosity(cartridge_cmd, project_without_dependencies, pack_format):
     assert all([log not in output for log in build_logs])
     assert 'Failed to run pre-build hook' in output
     assert prebuild_output in output
+
+
+@pytest.mark.parametrize('pack_format', ['deb', 'rpm'])
+def test_dependencies(cartridge_cmd, project_without_dependencies, pack_format, tarantool_versions, tmpdir):
+    project = project_without_dependencies
+
+    cmd = [
+        cartridge_cmd,
+        "pack", pack_format,
+        "--deps", "dependency01>=1.2,dependency01< 3",
+        "--deps", "dependency02 == 2.5 ",
+        "--deps", "\tdependency03    ",
+        project.path,
+    ]
+
+    if platform.system() == 'Darwin':
+        cmd.append('--use-docker')
+
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    assert rc == 0
+
+    if pack_format == 'deb':
+        deps = (
+            "dependency01 (>= 1.2)",
+            "dependency01 (<< 3)",
+            "dependency02 (= 2.5)",
+            "dependency03"
+        )
+
+        assert_dependencies_deb(find_archive(tmpdir, project.name, 'deb'), deps, tarantool_versions, tmpdir)
+    else:
+        deps = (
+            ("dependency01", 0x08 | 0x04, "1.2"),  # >=
+            ("dependency01", 0x02, "3"),  # <
+            ("dependency02", 0x08, "2.5"),  # =
+            ("dependency03", 0, "")
+        )
+
+        assert_dependencies_rpm(find_archive(tmpdir, project.name, 'rpm'), deps, tarantool_versions)
+
+
+@pytest.mark.parametrize('pack_format', ['deb', 'rpm'])
+def test_standard_dependencies_file(
+    cartridge_cmd, project_without_dependencies, pack_format, tarantool_versions, tmpdir
+):
+    project = project_without_dependencies
+
+    deps_filepath = os.path.join(tmpdir, "deps.txt")
+    with open(deps_filepath, "w") as f:
+        f.write("cool_dependency_01 >= 1.2, < 3 \n" +
+                "\t\n" +
+                " // comment\n" +
+                "\n" +
+                "  cool_dependency_02 == 2.5\n" +
+                "\tcool_dependency_03  ")
+
+    replace_project_file(project, 'package-deps.txt', deps_filepath)
+
+    cmd = [
+        cartridge_cmd,
+        "pack", pack_format,
+        project.path,
+    ]
+
+    if platform.system() == 'Darwin':
+        cmd.append('--use-docker')
+
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    assert rc == 0
+
+    if pack_format == 'deb':
+        deps = (
+            "cool_dependency_01 (>= 1.2)",
+            "cool_dependency_01 (<< 3)",
+            "cool_dependency_02 (= 2.5)",
+            "cool_dependency_03"
+        )
+
+        assert_dependencies_deb(find_archive(tmpdir, project.name, 'deb'), deps, tarantool_versions, tmpdir)
+    else:
+        deps = (
+            ("cool_dependency_01", 0x08 | 0x04, "1.2"),  # >=
+            ("cool_dependency_01", 0x02, "3"),  # <
+            ("cool_dependency_02", 0x08, "2.5"),  # =
+            ("cool_dependency_03", 0, "")
+        )
+
+        assert_dependencies_rpm(find_archive(tmpdir, project.name, 'rpm'), deps, tarantool_versions)
+
+
+@pytest.mark.parametrize('pack_format', ['deb', 'rpm'])
+def test_custom_dependencies_file(cartridge_cmd, project_without_dependencies, pack_format, tarantool_versions, tmpdir):
+    project = project_without_dependencies
+
+    deps_filepath = os.path.join(tmpdir, "deps.txt")
+    with open(deps_filepath, "w") as f:
+        f.write("dependency01 >= 1.2, < 3 \n" +
+                "  dependency02 == 2.5\n" +
+                "\tdependency03  ")
+
+    cmd = [
+        cartridge_cmd,
+        "pack", pack_format,
+        "--deps-file", deps_filepath,
+        project.path,
+    ]
+
+    if platform.system() == 'Darwin':
+        cmd.append('--use-docker')
+
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    assert rc == 0
+
+    if pack_format == 'deb':
+        deps = (
+            "dependency01 (>= 1.2)",
+            "dependency01 (<< 3)",
+            "dependency02 (= 2.5)",
+            "dependency03"
+        )
+
+        assert_dependencies_deb(find_archive(tmpdir, project.name, 'deb'), deps, tarantool_versions, tmpdir)
+    else:
+        deps = (
+            ("dependency01", 0x08 | 0x04, "1.2"),  # >=
+            ("dependency01", 0x02, "3"),  # <
+            ("dependency02", 0x08, "2.5"),  # =
+            ("dependency03", 0, "")
+        )
+
+        assert_dependencies_rpm(find_archive(tmpdir, project.name, 'rpm'), deps, tarantool_versions)
+
+
+@pytest.mark.parametrize('pack_format', ['docker', 'tgz'])
+def test_dependencies_not_rpm_deb(cartridge_cmd, project_without_dependencies, pack_format, tmpdir):
+    project = project_without_dependencies
+
+    cmd = [
+        cartridge_cmd,
+        "pack", pack_format,
+        "--deps-file", "dependencies.txt",
+        project.path,
+    ]
+
+    warning_message = "You specified the --deps-file flag, but you are not packaging RPM or DEB. Flag will be ignored"
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    assert rc == 0
+    assert warning_message in output
+
+    cmd = [
+        cartridge_cmd,
+        "pack", pack_format,
+        "--deps", "dependencies01",
+        project.path
+    ]
+
+    warning_message = "You specified the --deps flag, but you are not packaging RPM or DEB. Flag will be ignored"
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    assert rc == 0
+    assert warning_message in output
+
+
+@pytest.mark.parametrize('pack_format', ['deb', 'rpm'])
+def test_dependencies_same_flags(cartridge_cmd, project_without_dependencies, pack_format, tmpdir):
+    project = project_without_dependencies
+
+    deps_filepath = os.path.join(tmpdir, "deps.txt")
+    with open(deps_filepath, "w") as f:
+        f.write("")
+
+    cmd = [
+        cartridge_cmd,
+        "pack", pack_format,
+        "--deps-file", deps_filepath,
+        "--deps", "dependency01",
+        project.path,
+    ]
+
+    if platform.system() == 'Darwin':
+        cmd.append('--use-docker')
+
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    assert rc == 1
+    assert "You can't specify --deps and --deps-file flags at the same time" in output
+
+
+@pytest.mark.parametrize('pack_format', ['deb', 'rpm'])
+def test_dependencies_file_not_exist(cartridge_cmd, project_without_dependencies, pack_format, tmpdir):
+    project = project_without_dependencies
+
+    cmd = [
+        cartridge_cmd,
+        "pack", pack_format,
+        "--deps-file", "not_exist_file.txt",
+        project.path,
+    ]
+
+    if platform.system() == 'Darwin':
+        cmd.append('--use-docker')
+
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    assert rc == 1
+    assert "Invalid path to file with dependencies" in output
+
+
+@pytest.mark.parametrize('pack_format', ['deb', 'rpm'])
+def test_broken_dependencies(cartridge_cmd, project_without_dependencies, pack_format, tmpdir):
+    project = project_without_dependencies
+
+    broken_filepath = os.path.join(tmpdir, "broken.txt")
+    with open(broken_filepath, "w") as f:
+        f.write("dep01 >= 14, <= 25, > 14\n")
+
+    cmd = [
+        cartridge_cmd,
+        "pack", pack_format,
+        "--deps-file", broken_filepath,
+        project.path,
+    ]
+
+    if platform.system() == 'Darwin':
+        cmd.append('--use-docker')
+
+    error_message = "Failed to parse dependencies file: Error during parse dependencies file"
+    rc, output = run_command_and_get_output(cmd, cwd=tmpdir)
+    assert rc == 1
+    assert error_message in output